ref(alerts):  Soft deprecate projectalertruledetails get method
