+++ tests/sentry/incidents/endpoints/test_organization_alert_rule_details.py
@@ -107,6 +107,15 @@ class AlertRuleDetailsGetEndpointTest(AlertRuleDetailsBase, APITestCase):
 
         assert resp.data == serialize(self.alert_rule, serializer=DetailedAlertRuleSerializer())
 
+    def test_aggregate_translation(self):
+        self.create_team(organization=self.organization, members=[self.user])
+        self.login_as(self.user)
+        alert_rule = self.create_alert_rule(aggregate="count_unique(tags[sentry:user])")
+        with self.feature("organizations:incidents"):
+            resp = self.get_success_response(self.organization.slug, alert_rule.id)
+            assert resp.data["aggregate"] == "count_unique(user)"
+            assert alert_rule.snuba_query.aggregate == "count_unique(tags[sentry:user])"
+
     def test_expand_latest_incident(self):
         self.create_team(organization=self.organization, members=[self.user])
         self.login_as(self.user)
